Remove liveness from the faceID flow
